Move octree interfaces into its own file
