--- Source/HelixToolkit.SharpDX.Shared/Model/Scene/Abstract/ScreenSpacedNode.cs
@@ -1,19 +1,0 @@
-﻿/*
-The MIT License (MIT)
-Copyright (c) 2018 Helix Toolkit contributors
-*/
-using SharpDX;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-
-#if NETFX_CORE
-namespace HelixToolkit.UWP.Model.Scene
-#else
-namespace HelixToolkit.Wpf.SharpDX.Model.Scene
-#endif
-{
-    public class ScreenSpacedNode : GroupNode
-    {
-    }
-}
