+++ Source/HelixToolkit.SharpDX.Shared/Utilities/Octrees/StaticOctrees/StaticOctree.cs
@@ -17,93 +17,6 @@ namespace HelixToolkit.UWP
 namespace HelixToolkit.Wpf.SharpDX
 #endif
 {
-    /// <summary>
-    /// Interface for basic octree. Used to implement static octree and dynamic octree
-    /// </summary>
-    public interface IOctreeBasic
-    {
-        /// <summary>
-        /// Whether the tree has been built.
-        /// </summary>
-        bool TreeBuilt { get; }
-
-        /// <summary>
-        ///
-        /// </summary>
-        event EventHandler<EventArgs> OnHit;
-
-        /// <summary>
-        /// Output the hit path of the tree traverse. Only for debugging
-        /// </summary>
-        IList<BoundingBox> HitPathBoundingBoxes { get; }
-
-        /// <summary>
-        /// Octree parameter
-        /// </summary>
-        OctreeBuildParameter Parameter { get; }
-
-        /// <summary>
-        /// Gets the bound.
-        /// </summary>
-        /// <value>
-        /// The bound.
-        /// </value>
-        BoundingBox Bound { get; }
-
-        /// <summary>
-        /// Build the static octree
-        /// </summary>
-        void BuildTree();
-
-        /// <summary>
-        ///
-        /// </summary>
-        /// <param name="context"></param>
-        /// <param name="model"></param>
-        /// <param name="modelMatrix"></param>
-        /// <param name="rayWS"></param>
-        /// <param name="hits"></param>
-        /// <returns></returns>
-        bool HitTest(IRenderContext context, object model, Matrix modelMatrix, Ray rayWS, ref List<HitTestResult> hits);
-
-        /// <summary>
-        ///
-        /// </summary>
-        /// <param name="context"></param>
-        /// <param name="model"></param>
-        /// <param name="modelMatrix"></param>
-        /// <param name="rayWS"></param>
-        /// <param name="hits"></param>
-        /// <param name="hitThickness"></param>
-        /// <returns></returns>
-        bool HitTest(IRenderContext context, object model, Matrix modelMatrix, Ray rayWS, ref List<HitTestResult> hits, float hitThickness);
-
-        /// <summary>
-        ///
-        /// </summary>
-        /// <param name="context"></param>
-        /// <param name="point"></param>
-        /// <param name="results"></param>
-        /// <param name="heuristicSearchFactor"></param>
-        /// <returns></returns>
-        bool FindNearestPointFromPoint(IRenderContext context, ref Vector3 point, ref List<HitTestResult> results, float heuristicSearchFactor = 1f);
-
-        /// <summary>
-        ///
-        /// </summary>
-        /// <param name="context"></param>
-        /// <param name="sphere"></param>
-        /// <param name="points"></param>
-        /// <returns></returns>
-        bool FindNearestPointBySphere(IRenderContext context, ref BoundingSphere sphere, ref List<HitTestResult> points);
-
-        /// <summary>
-        /// Creates the octree line model for debugging or visualize the octree
-        /// </summary>
-        /// <returns></returns>
-        LineGeometry3D CreateOctreeLineModel();
-    }
-
     public abstract class StaticOctree<T> : IOctreeBasic
     {
         public const int OctantSize = 8;
